cleanup integ test
